GraphQL 오류 경계의 불필요한 wrapper를 제거한다
